# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VacancyMatch AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  message: string;
}

export interface Candidate {
  id: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  locationText?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  seniority?: string | null;
  skills: string[];
  titles: string[];
  industries: string[];
  /** @nullable */
  remotePreference?: string | null;
  /** @nullable */
  desiredSalaryMin?: number | null;
  /** @nullable */
  desiredSalaryMax?: number | null;
  /** @nullable */
  salaryCurrency?: string | null;
  /** @nullable */
  cvFileName?: string | null;
  /** @nullable */
  cvText?: string | null;
  status: string;
  /** @nullable */
  lastMatchedAt?: string | null;
  /** @nullable */
  bestMatchScore?: number | null;
  /** @nullable */
  matchCount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CandidateInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email?: string;
  phone?: string;
  currentTitle?: string;
  currentCompany?: string;
  locationText?: string;
  summary?: string;
  seniority?: string;
  skills?: string[];
  titles?: string[];
  industries?: string[];
  remotePreference?: string;
  desiredSalaryMin?: number;
  desiredSalaryMax?: number;
  salaryCurrency?: string;
}

export interface CandidateUpdate {
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  email?: string;
  phone?: string;
  currentTitle?: string;
  currentCompany?: string;
  locationText?: string;
  summary?: string;
  seniority?: string;
  skills?: string[];
  titles?: string[];
  industries?: string[];
  remotePreference?: string;
  desiredSalaryMin?: number;
  desiredSalaryMax?: number;
  salaryCurrency?: string;
  status?: string;
}

export interface CvUploadInput {
  /** @minLength 1 */
  fileName: string;
  /**
     * Extracted plain text of the CV (client extracts from PDF/DOCX/TXT)
     * @minLength 20
     */
  cvText: string;
}

export interface Job {
  id: string;
  /** @nullable */
  sourceId?: string | null;
  /** @nullable */
  sourceName?: string | null;
  title: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  locationText?: string | null;
  /** @nullable */
  remoteType?: string | null;
  /** @nullable */
  employmentType?: string | null;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  salaryCurrency?: string | null;
  /** @nullable */
  descriptionText?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  industry?: string | null;
  skills: string[];
  /** @nullable */
  postedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  status: string;
  createdAt: string;
}

export interface JobSource {
  id: string;
  name: string;
  sourceType: string;
  /** @nullable */
  baseUrl?: string | null;
  isActive: boolean;
  /** @nullable */
  healthStatus?: string | null;
  /** @nullable */
  lastSyncAt?: string | null;
  jobCount: number;
  createdAt: string;
}

export interface JobSourceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sourceType: string;
  baseUrl?: string;
  isActive?: boolean;
}

export interface JobSourceUpdate {
  /** @minLength 1 */
  name?: string;
  sourceType?: string;
  baseUrl?: string;
  isActive?: boolean;
}

export interface ScoreBreakdown {
  skills: number;
  title: number;
  industry: number;
  location: number;
  comp: number;
  recency: number;
}

export interface Match {
  id: string;
  matchRunId: string;
  candidateId: string;
  /** @nullable */
  candidateName?: string | null;
  jobId: string;
  job?: Job;
  overallScore: number;
  scoreBreakdown: ScoreBreakdown;
  explanation: string[];
  matchedSkills: string[];
  missingSkills?: string[];
  recruiterStatus: string;
  /** @nullable */
  recruiterNote?: string | null;
  pushedToCrm: boolean;
  createdAt: string;
}

export interface MatchUpdate {
  /** new | shortlisted | dismissed | pushed */
  recruiterStatus?: string;
  recruiterNote?: string;
}

export interface MatchRun {
  id: string;
  candidateId: string;
  triggerType: string;
  /** @nullable */
  modelVersion?: string | null;
  status: string;
  matchCount: number;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface CrmPushInput {
  /** @minLength 1 */
  crmName: string;
  note?: string;
}

export interface CrmSyncEvent {
  id: string;
  /** @nullable */
  candidateId?: string | null;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  jobId?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  direction: string;
  crmName: string;
  /** @nullable */
  payloadSummary?: string | null;
  status: string;
  createdAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface AlertRule {
  id: string;
  candidateId: string;
  /** @nullable */
  candidateName?: string | null;
  minScore: number;
  frequency: string;
  isActive: boolean;
  /** @nullable */
  lastCheckedAt?: string | null;
  /** @nullable */
  lastTriggeredAt?: string | null;
  createdAt: string;
}

export interface AlertRuleInput {
  /** @minLength 1 */
  candidateId: string;
  minScore: number;
  /** @minLength 1 */
  frequency: string;
  isActive?: boolean;
}

export interface AlertRuleUpdate {
  minScore?: number;
  frequency?: string;
  isActive?: boolean;
}

export interface DashboardSummary {
  candidateCount: number;
  activeJobCount: number;
  activeSourceCount: number;
  matchRunCount: number;
  shortlistedCount: number;
  pushedToCrmCount: number;
  avgTopScore: number;
  newMatchesThisWeek: number;
}

export interface ActivityItem {
  id: string;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  description: string;
  createdAt: string;
}

export interface AuditLog {
  id: string;
  actorName: string;
  action: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  metadata?: string | null;
  createdAt: string;
}

export type ListCandidatesParams = {
search?: string;
status?: string;
};

export type ListMatchesParams = {
recruiterStatus?: string;
minScore?: number;
};

export type ListJobsParams = {
search?: string;
sourceId?: string;
status?: string;
};