# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VacancyMatch AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List all matches across candidates
      parameters:
        - name: recruiterStatus
          in: query
          required: false
          schema: { type: string }
        - name: minScore
          in: query
          required: false
          schema: { type: number }
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /matches/{id}:
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update recruiter decision on a match (shortlist, dismiss, note)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /matches/{id}/push-to-crm:
    post:
      operationId: pushMatchToCrm
      tags: [matches]
      summary: Push a match to the connected CRM
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmPushInput"
      responses:
        "201":
          description: CRM sync event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmSyncEvent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Search normalized live jobs
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: sourceId
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Job:
      type: object
      required: [id, title, status, skills, createdAt]
      properties:
        id: { type: string }
        sourceId: { type: ["string", "null"] }
        sourceName: { type: ["string", "null"] }
        title: { type: string }
        companyName: { type: ["string", "null"] }
        locationText: { type: ["string", "null"] }
        remoteType: { type: ["string", "null"] }
        employmentType: { type: ["string", "null"] }
        salaryMin: { type: ["number", "null"] }
        salaryMax: { type: ["number", "null"] }
        salaryCurrency: { type: ["string", "null"] }
        descriptionText: { type: ["string", "null"] }
        applyUrl: { type: ["string", "null"] }
        industry: { type: ["string", "null"] }
        skills:
          type: array
          items: { type: string }
        postedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
...
    ScoreBreakdown:
      type: object
      required: [skills, title, industry, location, comp, recency]
      properties:
        skills: { type: number }
        title: { type: number }
        industry: { type: number }
        location: { type: number }
        comp: { type: number }
        recency: { type: number }
    Match:
      type: object
      required:
        [
          id,
          matchRunId,
          candidateId,
          jobId,
          overallScore,
          scoreBreakdown,
          explanation,
          matchedSkills,
          recruiterStatus,
          pushedToCrm,
          createdAt,
        ]
      properties:
        id: { type: string }
        matchRunId: { type: string }
        candidateId: { type: string }
        candidateName: { type: ["string", "null"] }
        jobId: { type: string }
        job:
          $ref: "#/components/schemas/Job"
        overallScore: { type: number }
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        explanation:
          type: array
          items: { type: string }
        matchedSkills:
          type: array
          items: { type: string }
        missingSkills:
          type: array
          items: { type: string }
        recruiterStatus: { type: string }
        recruiterNote: { type: ["string", "null"] }
        pushedToCrm: { type: boolean }
        createdAt: { type: string }
    MatchUpdate:
      type: object
      properties:
        recruiterStatus:
          type: string
          description: new | shortlisted | dismissed | pushed
        recruiterNote: { type: string }
...
    CrmPushInput:
      type: object
      required: [crmName]
      properties:
        crmName: { type: string, minLength: 1 }
        note: { type: string }
    CrmSyncEvent:
      type: object
      required: [id, direction, crmName, status, createdAt]
      properties:
        id: { type: string }
        candidateId: { type: ["string", "null"] }
        candidateName: { type: ["string", "null"] }
        jobId: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        direction: { type: string }
        crmName: { type: string }
        payloadSummary: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        processedAt: { type: ["string", "null"] }
...