# lib/api-zod/src/generated/types/matchUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VacancyMatch AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchUpdate {
  /** new | shortlisted | dismissed | pushed */
  recruiterStatus?: string;
  recruiterNote?: string;
}